Client: Refactor client
The following change firstly bumps juju/http to v2, as this contains a
lot of breaking changes.

Secondly it introduces the concept of transport middleware. In order to
add additional middlewares in the future (thinking about retry and
metric transport wrappers), then we need to expose this to the consumer
of the library.

Part of that standardisation was to correctly setup a default transport
so that all subsquent transports use the same one. Instead of some using
a blank transport and others using a custom one.

The middleware concept then allows us to talk about concepts like
breakers, that prevent outgoing calls at a much higher level.

Lastly we moved from a fat config, to providing options that can be
sequenced togther, before passing it to the constructor. One benefit of
doing it this way is allowing us to create defaults in one location and
pass that through, rather than ad-hoc inline style.