Add client option to record outgoing requests
The request recorder wraps an http.RoundTripper and invokes a set of
provided Record methods depending on the outcome of the request.

If the wrapped RoundTripper returns back without an error, the Record
method is invoked with the following arguments: request method, request
URL, http.Response instance and request roundtrip time (RTT).

On the other hand, if the RoundTripper returned back an error (and a nil
response), the RecordError is invoked with the following arguments:
request method, request URL, roundtripper error.

One or more RequestRecoder instances can be registered when creating a
new http client using the WithRequestRecorder() option.