Middleware: Introduce a retry roundtripper
This is the beginnings of the retry transport. The idea being that we
allow global retries depending on the errors passed in.

This isn't smart about recommencing already downloaded content, this is
too low down to be able to do that, instead this is a generic attempt at
solving the problem where network stacks are flakey.