Middleware: handle retries with dates
According to the RFC7231 spec it states that the Retry-After header
could actually be either a delay in seconds or a date in the future. We
only previously handled delays in seconds and that would have failed
because we just didn't correctly handle it.

The fix was to handle that and also so clamp it to a maximum duration as
we would never want to wait for too long as that could potentially dead
lock the running program.