Do not allow http pkg to cache proxy settings.
Fix regression introduce in refactoring this package, LP1940327.
http.ProxyFromEnvironment will cache the proxy settings for the life of
the using process. This is unfortunate as juju allows user to change the
proxy settings on the fly. The trade off is more calls to find the
environment values for proxy settings.